topics combobox, ability to add topics
