+++ ui/src/components/Publication/Creation/Links/LinkedPublicationRow.tsx
